@@ -1,76 +1,58 @@
 {
   "name": "rail-rakshak-frontend",
-  "version": "1.0.0",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@chakra-ui/hooks": "^2.1.4",
-    "@chakra-ui/modal": "^2.2.9",
-    "@chakra-ui/popover": "^2.1.8",
-    "@chakra-ui/portal": "^2.0.15",
-    "@chakra-ui/system": "^2.3.5",
-    "@chakra-ui/tooltip": "^2.2.6",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
+    "@chakra-ui/hooks": "^2.2.1",
+    "@chakra-ui/modal": "^2.3.1",
+    "@chakra-ui/popover": "^2.2.1",
+    "@chakra-ui/portal": "^2.1.0",
+    "@chakra-ui/system": "^2.6.1",
+    "@chakra-ui/tooltip": "^2.3.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@reduxjs/toolkit": "^1.9.5",
     "@tanstack/match-sorter-utils": "^8.8.4",
-    "@tanstack/react-table": "^8.7.9",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^25.2.3",
-    "@types/node": "^12.20.55",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "apexcharts": "^3.35.5",
+    "@tanstack/react-table": "^8.10.1",
+    "@types/history": "^5.0.0",
+    "apexcharts": "^3.42.0",
     "axios": "^1.5.0",
-    "framer-motion": "^7.10.2",
+    "framer-motion": "^10.16.4",
     "react": "^18.2.0",
-    "react-apexcharts": "^1.4.0",
-    "react-calendar": "^3.9.0",
+    "react-apexcharts": "^1.4.1",
+    "react-calendar": "^4.6.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.4.0",
+    "react-icons": "^4.11.0",
     "react-redux": "^8.1.2",
-    "react-router-dom": "^6.4.0",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.16.0",
     "react-table": "^7.8.0",
     "react-use-websocket": "^4.4.0",
-    "typescript": "^4.7.4",
-    "video.js": "^8.5.2",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "pretty": "prettier --write \"./**/*.{js,jsx,json}\""
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "video.js": "^8.5.2"
   },
   "devDependencies": {
-    "@types/react-calendar": "^3.5.2",
+    "@types/node": "^20.6.4",
+    "@types/react": "^18.2.18",
+    "@types/react-calendar": "^4.1.0",
+    "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-table": "^7.7.12",
-    "autoprefixer": "^10.4.8",
-    "postcss": "^8.4.16",
-    "prettier": "^2.8.3",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "tailwindcss": "^3.1.8"
+    "@types/react-table": "^7.7.15",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
+    "@vitejs/plugin-react": "^4.0.4",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.46.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "postcss": "^8.4.30",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.8"
   }
 }
