@@ -1,8 +1,8 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 
-import AdminLayout from "layouts/admin";
-import AuthLayout from "layouts/auth";
-import ProtectedRoute from "layouts/ProtectedRoute";
+import AdminLayout from "./layouts/admin";
+import AuthLayout from "./layouts/auth";
+
 const App = () => {
   return (
     <Routes>
