@@ -1,9 +1,20 @@
-import { PayloadAction, createSlice } from "@reduxjs/toolkit";
+import { PayloadAction, createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import axios from "axios";
+import { BACKEND_URL } from "@/constants/definitions";
 
 export interface Task {
   id: string;
   title: string;
-  desc: string;
+  description: string;
+  assigned_to: string[];
+  image: string;
+  created_at: string;
+  deadline: string;
+  type: string;
+  status: string;
+  assc_incident: string;
+  dept_name: string;
+  station_name: string;
 }
 
 interface TaskState {
@@ -18,6 +29,37 @@ const initialState: TaskState = {
   error: null,
 };
 
+export const fetchTasks = createAsyncThunk(
+  "task/fetch",
+  async (payload: { deptName: string; stationName: string }, thunkAPI) => {
+    try {
+      const { deptName, stationName } = payload;
+      const res = await axios.get(
+        `${BACKEND_URL}/tasks/get_tasks_by_dept?dept_name=${deptName}&station_name=${stationName}`
+      );
+      console.log(res.data);
+      return res.data;
+    } catch (error) {
+      console.log(error);
+      throw error;
+    }
+  }
+);
+
+export const addTask = createAsyncThunk(
+  "task/add",
+  async (payload: Task, thunkAPI) => {
+    try {
+      const res = await axios.post(`${BACKEND_URL}/tasks/create_task`, payload);
+      console.log(res.data);
+      return res.data;
+    } catch (error) {
+      console.log(error);
+      throw error;
+    }
+  }
+);
+
 export const TaskSlice = createSlice({
   name: "tasks",
   initialState,
@@ -30,6 +72,34 @@ export const TaskSlice = createSlice({
       state.data = [];
       state.error = null;
     },
+  },
+  extraReducers: (builder) => {
+    builder.addCase(fetchTasks.pending, (state) => {
+      state.loading = true;
+      state.error = null;
+    });
+    builder.addCase(fetchTasks.fulfilled, (state, action) => {
+      state.data = action.payload.SUCCESS;
+      state.loading = false;
+      state.error = null;
+    });
+    builder.addCase(fetchTasks.rejected, (state, action) => {
+      state.loading = false;
+      state.error = action.error.message || null;
+    });
+    builder.addCase(addTask.pending, (state) => {
+      state.loading = true;
+      state.error = null;
+    });
+    builder.addCase(addTask.fulfilled, (state, action) => {
+      state.data = [...state.data, action.payload.SUCCESS];
+      state.loading = false;
+      state.error = null;
+    });
+    builder.addCase(addTask.rejected, (state, action) => {
+      state.loading = false;
+      state.error = action.error.message || null;
+    });
   },
 });
 
