@@ -1,6 +1,6 @@
 import { IoHeart, IoHeartOutline } from "react-icons/io5";
 import { useState } from "react";
-import Card from "components/card";
+import Card from "@/components/card";
 
 const NftCard = (props: {
   image: string;
