@@ -1,6 +1,6 @@
 /* eslint-disable */
 import { Link, useLocation } from "react-router-dom";
-import DashIcon from "components/icons/DashIcon";
+import DashIcon from "@/components/icons/DashIcon";
 // chakra imports
 
 export const SidebarLinks = (props: { routes: RoutesType[] }): JSX.Element => {
