@@ -1,10 +1,10 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import "./index.css";
-
-import App from "./App";
 import { Provider } from "react-redux";
-import { store } from "app/store";
+
+import { store } from "./app/store.ts";
+import App from "./App.tsx";
+import "./index.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
