@@ -1,5 +1,5 @@
 import { useState } from "react";
-import Card from "components/card";
+import Card from "@/components/card";
 
 import {
   createColumnHelper,
